Allow swipe back to home view on top of graph scroll view #14
Added KSTPanningScrollView, which allows its delegate to implement a custom action instead of bounding if the user is already on an edge of the scroll view and attempts to scroll more in that direction.
Made KSTViewController a KSTPanningScrollViewDelegate so it can pan back to the home screen when the user swipes left at the leftmost side of the graph.